Move FewerTags class to separate file
